feat: split coins tracking routes
