+++ src/app/coins/layout.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+export default function CoinsLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
+}
