+++ src/app/coins/map/page.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function CoinsMapPage() {
+  return null;
+}
