+++ src/app/coins/page.tsx
@@ -1,194 +1,14 @@
 "use client";
 
-import React, { useEffect, useMemo, useState } from "react";
-
-import { CheckBadgeIcon } from "@heroicons/react/16/solid";
-import coins from "@/data/coins.json";
-import { useSearchParams } from "next/navigation";
+import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import Filter from "@/components/Filter";
-import { randomId } from "@/utils";
 
 export default function CoinsPage() {
   const router = useRouter();
-  const searchParams = useSearchParams();
-
-  // Get initial filter values from URL query params or default to "all"
-  const initialLocationFilter = searchParams.get("filters[location]") || "all";
-
-  const initialCollectedFilter =
-    searchParams.get("filters[collected]") || "all";
-
-  const [locationFilter, setLocationFilter] = useState<string>(
-    initialLocationFilter
-  );
-
-  const [collectedFilter, setCollectedFilter] = useState<string>(
-    initialCollectedFilter
-  );
-
-  const locationsByProvince = useMemo(() => {
-    const provinces = [...new Set(coins.map((coin) => coin.province))].sort();
-
-    const result = provinces.map((province) => {
-      const citiesInProvince = [
-        ...new Set(
-          coins
-            .filter((coin) => coin.province === province)
-            .map((coin) => coin.location)
-            .sort()
-        ),
-      ];
-
-      return {
-        value: province,
-        label: province,
-        cities: citiesInProvince.map((cityName) => {
-          return {
-            value: cityName,
-            label: cityName,
-          };
-        }),
-      };
-    });
-
-    return result;
-  }, []);
-
-  const locations = useMemo(() => {
-    const options = [
-      {
-        key: randomId(),
-        value: "all",
-        label: "Всички",
-      },
-    ];
-
-    locationsByProvince.forEach((province) => {
-      options.push({
-        key: `${randomId()}-${province.value}`,
-        value: province.value,
-        label: province.label,
-      });
-
-      province.cities.forEach((city) => {
-        options.push({
-          key: `${randomId()}-${province.value}-${city.value}`,
-          value: city.value,
-          label: `- ${city.label}`,
-        });
-      });
-    });
-
-    return options;
-  }, [locationsByProvince]);
-
-  const collectedFilters = [
-    { value: "all", label: "Всички" },
-    { value: "yes", label: "Да" },
-    { value: "no", label: "Не" },
-  ];
-
-  const onLocationFilterChange = (value: string) => {
-    setLocationFilter(value);
-  };
-
-  const onCollectedFilterChange = (value: string) => {
-    setCollectedFilter(value);
-  };
-
-  const filteredData = coins.filter((coin) => {
-    let passesLocationFilter = true;
-    let passesCollectedFilter = true;
-
-    if (locationFilter !== "all") {
-      passesLocationFilter =
-        coin.province === locationFilter || coin.location === locationFilter;
-    }
-
-    if (collectedFilter !== "all") {
-      if (collectedFilter === "yes") {
-        passesCollectedFilter = coin.collected === true;
-      } else if (collectedFilter === "no") {
-        passesCollectedFilter = coin.collected === false;
-      }
-    }
-
-    return passesLocationFilter && passesCollectedFilter;
-  });
 
   useEffect(() => {
-    const queryParts = [
-      `filters[location]=${encodeURIComponent(locationFilter)}`,
-      `filters[collected]=${encodeURIComponent(collectedFilter)}`,
-    ];
+    router.replace("/coins/list");
+  }, [router]);
 
-    router.replace(`/coins?${queryParts.join("&")}`);
-  }, [locationFilter, collectedFilter, router]);
-
-  return (
-    <>
-      <section aria-labelledby="filter-heading" className="mx-auto  py-6">
-        <div className="flex flex-col md:flex-row md:items-center md:justify-between">
-          <div className="flex items-center gap-x-5">
-            <Filter
-              name="Локация"
-              selectedValue={locationFilter}
-              options={locations}
-              onFilterChanged={onLocationFilterChange}
-            />
-
-            <Filter
-              name="Събрани"
-              selectedValue={collectedFilter}
-              options={collectedFilters}
-              onFilterChanged={onCollectedFilterChange}
-            />
-          </div>
-
-          <div className="text-sm italic pt-5 md:pt-0">
-            {filteredData.length} резултата от общо {coins.length}
-          </div>
-        </div>
-      </section>
-
-      {filteredData.length > 0 && (
-        <>
-          <ul
-            role="list"
-            className="grid grid-cols-2 gap-6 md:grid-cols-3 lg:grid-cols-4"
-          >
-            {filteredData.map((coin) => (
-              <li
-                key={coin.id}
-                className="col-span-1 flex flex-col relative divide-y divide-gray-200 rounded-lg bg-white text-center shadow-sm"
-              >
-                <div className="flex flex-1 flex-col p-8">
-                  <a href={coin.url} target="_blank" rel="noopener noreferrer">
-                    <span aria-hidden="true" className="absolute inset-0" />
-                    <img
-                      alt={`Coin image for ${coin.name}`}
-                      src={coin.images[0].url}
-                      className="mx-auto size-32 shrink-0 rounded-full"
-                    />
-                  </a>
-                  <h3 className="mt-6 text-sm font-medium text-gray-900">
-                    {coin.name}
-                  </h3>
-                  {coin.collected && (
-                    <div className="mt-3 flex flex-col items-center">
-                      <CheckBadgeIcon
-                        title="Събрана монета"
-                        className="w-5 h-5 text-green-700"
-                      />
-                    </div>
-                  )}
-                </div>
-              </li>
-            ))}
-          </ul>
-        </>
-      )}
-    </>
-  );
+  return null;
 }
